rg-220: * dropdown options to a separate file
