+++ frontend/src/bundles/edit-template/constants/constants.ts
@@ -1,0 +1,1 @@
+export { dropdownOptions } from './dropdown-items.constant';
