+++ frontend/src/bundles/edit-template/constants/dropdown-items.constant.ts
@@ -1,0 +1,22 @@
+import { FontStyles } from '../enums/font-styles';
+
+const dropdownOptions = [
+    {
+        label: FontStyles.Regular,
+        value: FontStyles.Regular,
+    },
+    {
+        label: FontStyles.Bold,
+        value: FontStyles.Bold,
+    },
+    {
+        label: FontStyles.Italic,
+        value: FontStyles.Italic,
+    },
+    {
+        label: FontStyles.BoldItalic,
+        value: FontStyles.BoldItalic,
+    },
+];
+
+export { dropdownOptions };
